crypto: rename test directory to tests
Tooling like the includes linter assumes that test directories are named
as plural 'tests' which fits with naming convention of a large majority
of existing modules in the tree. Renaming the paths to avoid more
special casing in tooling.
